backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-ai
google-generativeai
python-dotenv
pgvector
alembic
passlib[bcrypt]
python-jose[cryptography]
email-validator
